requirements.txt
psycopg2-binary==2.9.1
Django==3.2.5
djangorestframework==3.12.4
django-registration==3.2
